moved code related to "accounts" updated reference links
