+++ frontend/src/features/accounts/types/index.ts
@@ -1,0 +1,20 @@
+export interface TwitterUser {
+	id: bigint;
+	name: string;
+	screenName: string;
+	profileImageUrl: string;
+}
+
+export interface Account {
+	id: number;
+	userId: number;
+	twitterUser: TwitterUser;
+}
+
+export type Accounts = Array<Account>;
+
+export interface AccountContextProps {
+	account: Account;
+	setAccount?: (account: Account) => void;
+	handleAccountChange: (account: Account) => void;
+}
