+++ frontend/src/types/index.ts
@@ -1,44 +1,3 @@
-export interface TwitterUser {
-	id: bigint;
-	name: string;
-	screenName: string;
-	profileImageUrl: string;
-}
-
-export interface Account {
-	id: number;
-	userId: number;
-	twitterUser: TwitterUser;
-}
-
-export type Accounts = Array<Account>;
-
-// export interface TwitterAccount {
-// 	userId: string;
-// 	name: string;
-// 	screenName: string;
-// 	profileImageUrl: string;
-// }
-
-export interface AccountContextProps {
-	account: Account;
-	setAccount?: (account: Account) => void;
-	handleAccountChange: (account: Account) => void;
-}
-
-// export type TwitterAccounts = Array<TwitterAccount>;
-
-// Single Tweet
-// export interface Tweet {
-// 	id?: number;
-// 	text: string;
-// 	image64: string;
-// 	paperId: number;
-// 	userId?: string;
-// 	scheduledTimeUTC?: string;
-// 	sent?: boolean;
-// }
-
 export interface Tweet {
 	id: number;
 	twitterUserId: bigint;
